+++ tests/sokolnikov911/YandexTurboPages/RelatedItemsListTest.php
@@ -30,6 +30,13 @@ class RelatedItemsListTest extends TestCase
         $this->assertXmlStringEqualsXmlString($expect, $relatedItemsList->asXML()->asXML());
     }
 
+    public function testAsXMLWithInfinity()
+    {
+        $relatedItemsList = new RelatedItemsList(true);
+        $expect = '<yandex:related type="infinity"/>';
+        $this->assertXmlStringEqualsXmlString($expect, $relatedItemsList->asXML()->asXML());
+    }
+
     public function testAsXMLWithRelatedPage()
     {
         $relatedItemsList = new RelatedItemsList();